+++ backend/canisters/community/impl/src/updates/send_message.rs
@@ -5,7 +5,6 @@ use crate::timer_job_types::{DeleteFileReferencesJob, EndPollJob, MarkP2PSwapExp
 use crate::{mutate_state, run_regular_jobs, Data, RuntimeState};
 use canister_api_macros::update;
 use canister_tracing_macros::trace;
-use chat_events::OPENCHAT_BOT_USER_ID;
 use community_canister::c2c_send_message::{Args as C2CArgs, Response as C2CResponse};
 use community_canister::send_message::{Response::*, *};
 use group_chat_core::SendMessageResult;
@@ -18,6 +17,7 @@ use types::{
     Notification, TimestampMillis, User, UserId, UserType, Version,
 };
 use user_canister::{CommunityCanisterEvent, MessageActivity, MessageActivityEvent};
+use utils::consts::OPENCHAT_BOT_USER_ID;
 
 #[update(candid = true, msgpack = true)]
 #[trace]